requirements.txt
altgraph==0.17.2
numpy==1.21.3
Pillow==8.4.0
pycairo==1.20.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
